25/07/2025: Apenas algumas mudanças
Resultado da MCS da F7 sem FPU;
Começo da MCS da F7 em double com FPU;
Aplicação em Python que mostra em tempo real a posição do acelerômetro

+++ CubeIDE/MTS_certa/Core/Inc/calib.h
@@ -13,11 +13,11 @@
 #define N 9
 #define tam 1112
 
-//extern float Ht_H[N][N], Ht_e[N], inv[N][N], mul_mat[N][tam];
+//extern double Ht_H[N][N], Ht_e[N], inv[N][N], mul_mat[N][tam];
 
-void ETS(float *mx, float *my, float *mz, float *p);
-uint8_t NLLS(float *mx, float *my, float *mz, float *p);
+void ETS(double *mx, double *my, double *mz, double *p);
+uint8_t NLLS(double *mx, double *my, double *mz, double *p);
 
-void inverter_matriz(float matriz[N][N], float inversa[N][N]);
+void inverter_matriz(double matriz[N][N], double inversa[N][N]);
 
 #endif /* INC_CALIB_H_ */
